pyproject.toml
[project]
name = "image-gen-with-bria"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.3",
    "django-allauth>=65.10.0",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "requests>=2.32.4",
    "typing>=3.10.0.0",
]
